analytics/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
mysql-connector-python==8.0.26
pymongo==4.0.1